Refactor apply_pca function to use df.update and optimize column addition
This commit refactors the apply_pca function in the Data.py file. Instead of looping through columns to update the dataframe, the function now uses the df.update method for better performance. Additionally, the code has been optimized to add columns from the combined DataFrame to the original DataFrame using a for loop. This change improves the efficiency of the function.

Fix potential issue with loading data in load_data function

This commit fixes a potential issue in the load_data function in the Data.py file. The columns "experimental" has been added to the list of columns to be dropped when splitting the PDBbind data into training and testing sets. This change ensures that the "experimental" column is ignored during the split process.